allow create enroll to finance

+++ routes/adminRoute.js
@@ -3,14 +3,14 @@ const express = require("express");
 
 const { protect } = require("../middlewares/authMiddleware.js");
 const { authorize } = require("../middlewares/roleMiddleware.js");
-const { getAllUsers, getUserById, updateUser, changeUserPassword, deleteUserById, deleteAllUsers, createUser, assignRole } = require("../controllers/adminController.js");
+const { getAllUsers, getUserById, updateUser, changeUserPassword, deleteUserById, deleteAllUsers, createUser, assignRole, getAdminRecipients } = require("../controllers/adminController.js");
 
 const router = express.Router();
 
 
 // 🔒 Only Admin Can Access All Routes
 
-router.get("/users", protect, authorize("admin", "super_admin", "sales_manager"), getAllUsers);
+router.get("/users", protect, authorize("admin", "super_admin", "sales_manager", "finance_manager"), getAllUsers);
 router.get("/users/:id", protect, authorize("admin", "super_admin"), getUserById);
 router.patch("/users/:id", protect, authorize("admin", "super_admin"), updateUser);
 router.patch("/users/:id/change-password", protect, authorize("admin", "super_admin"), changeUserPassword);
@@ -18,5 +18,6 @@ router.delete("/users/:id", protect, authorize("admin", "super_admin"), deleteUs
 router.delete("/users", protect, authorize("admin", "super_admin"), deleteAllUsers);
 router.post("/users", protect, authorize("admin", "super_admin"), createUser);
 router.patch("/users/:id/role", protect, authorize("admin", "super_admin", "sales_manager"), assignRole);
+router.get("/users/recipients", protect, authorize("admin", "super_admin"), getAdminRecipients);
 
 module.exports = router;
